correcciones

+++ database/migrations/2025_08_24_174819_create_inventarios_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nombre_producto');
             $table->string('descripcion')->nullable();
-            $table->integer('precio_unitario')->default(0);
+            $table->decimal('precio_unitario', 8, 2)->default(0);
             $table->boolean('sin_stock')->default(false);
             $table->integer('stock')->default(0);
             $table->unsignedBigInteger('categoria_id');
